requirements.txt
# Core dependencies
pyspark==3.5.0
delta-spark==3.2.0
pandas==2.2.3
PyYAML==6.0.2

# Testing dependencies
pytest==8.3.3

# Linting and code formatting
ruff==0.6.9